Ai-Track-Mohith-main/server/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic[email]>=2.7
pydantic-settings>=2.3
sqlalchemy>=2.0
asyncpg>=0.30
bcrypt>=4.0
pyjwt[crypto]>=2.8
python-multipart>=0.0.9
alembic>=1.14
loguru>=0.7
google-adk>=1.0
python-dotenv>=1.0
